Move interface, struct and checkwebsite func to service package
